[1127] TMU 序列化迁移到 moebius
将 from_tmu.cpp + to_tmu.cpp 合并为 moebius/Data/Convert/tmu.cpp,
声明收入 tmu.hpp。get_codes 连同其 helper 内联为 static,is_snippet
收口到 moebius tree_helper,主工程 convert.hpp 改为转发。删除
src/Data/Convert/Mogan/ 下的两个旧文件,顺手清理 from_tmu 顶部
无用的 preferences.hpp include。
