+++ src/Data/Convert/Generic/generic.cpp
@@ -16,15 +16,6 @@
 
 static url current_file_focus= url_none ();
 
-bool
-is_snippet (tree doc) {
-  if (!is_document (doc)) return true;
-  int i, n= N (doc);
-  for (i= 0; i < n; i++)
-    if (is_compound (doc[i], "TeXmacs", 1)) return false;
-  return true;
-}
-
 url
 get_file_focus () {
   return current_file_focus;
